tests: Protect all preloaded functions against recursion

+++ tests/socket-add-extra-source.c
@@ -17,48 +17,55 @@ static int extra_fd = -1;
 // When dtee connects to its own input socket, clone the socket details and make an extra socket
 int connect(int sockfd, const struct sockaddr *addr, socklen_t addrlen) {
 	int (*next_connect)(int, const struct sockaddr *, socklen_t) = dlsym(RTLD_NEXT, "connect");
+	static __thread bool active = false;
 
-	if (dtee_test_is_dtee() && extra_fd < 0) {
-		struct sockaddr_un sock_addr;
+	if (!active) {
+		active = true;
 
-		if (dtee_test_is_fd_unix_socket(sockfd, &sock_addr)) {
-			char path[sizeof(sock_addr.sun_path) + 1] = { 0 };
+		if (dtee_test_is_dtee() && extra_fd < 0) {
+			struct sockaddr_un sock_addr;
 
-			memcpy(path, &sock_addr.sun_path, sizeof(sock_addr.sun_path));
-			path[strlen(path) - 1] = 'x';
-			strncpy(sock_addr.sun_path, path, sizeof(sock_addr.sun_path) - 1);
+			if (dtee_test_is_fd_unix_socket(sockfd, &sock_addr)) {
+				char path[sizeof(sock_addr.sun_path) + 1] = { 0 };
 
-			extra_fd = socket(AF_UNIX, SOCK_DGRAM, 0);
-			if (extra_fd >= 0) {
-				if (!bind(extra_fd, (struct sockaddr*)&sock_addr, sizeof(sock_addr))) {
-					if (!connect(extra_fd, addr, addrlen)) {
-						printf("socket-add-extra-source: created socket \"%s\" connected to \"%s\"\n",
-							sock_addr.sun_path, path);
-						fflush(stdout);
+				memcpy(path, &sock_addr.sun_path, sizeof(sock_addr.sun_path));
+				path[strlen(path) - 1] = 'x';
+				strncpy(sock_addr.sun_path, path, sizeof(sock_addr.sun_path) - 1);
+
+				extra_fd = socket(AF_UNIX, SOCK_DGRAM, 0);
+				if (extra_fd >= 0) {
+					if (!bind(extra_fd, (struct sockaddr*)&sock_addr, sizeof(sock_addr))) {
+						if (!connect(extra_fd, addr, addrlen)) {
+							printf("socket-add-extra-source: created socket \"%s\" connected to \"%s\"\n",
+								sock_addr.sun_path, path);
+							fflush(stdout);
+						} else {
+							perror("socket-add-extra-source: connect");
+							fflush(stderr);
+
+							close(extra_fd);
+							extra_fd = -1;
+						}
+
+						if (unlink(path)) {
+							perror("socket-add-extra-source: unlink");
+							fflush(stderr);
+						}
 					} else {
-						perror("socket-add-extra-source: connect");
+						perror("socket-add-extra-source: bind");
 						fflush(stderr);
 
 						close(extra_fd);
 						extra_fd = -1;
 					}
-
-					if (unlink(path)) {
-						perror("socket-add-extra-source: unlink");
-						fflush(stderr);
-					}
 				} else {
-					perror("socket-add-extra-source: bind");
+					perror("socket-add-extra-source: socket");
 					fflush(stderr);
-
-					close(extra_fd);
-					extra_fd = -1;
 				}
-			} else {
-				perror("socket-add-extra-source: socket");
-				fflush(stderr);
 			}
 		}
+
+		active = false;
 	}
 
 	return (*next_connect)(sockfd, addr, addrlen);
@@ -67,26 +74,33 @@ int connect(int sockfd, const struct sockaddr *addr, socklen_t addrlen) {
 // When dtee is about to exec the command, insert an additional message
 int execvp(const char *filename, char *const argv[]) {
 	int (*next_execvp)(const char*, char * const *) = dlsym(RTLD_NEXT, "execvp");
+	static __thread bool active = false;
 
-	if (dtee_test_is_dtee()) {
-		if (extra_fd >= 0) {
-			const char *message = "Message from unknown source\n";
+	if (!active) {
+		active = true;
 
-			printf("socket-add-extra-source: inserting additional message\n");
-			fflush(stdout);
+		if (dtee_test_is_dtee()) {
+			if (extra_fd >= 0) {
+				const char *message = "Message from unknown source\n";
 
-			int len = strlen(message);
-			if (write(extra_fd, message, strlen(message)) != len) {
-				perror("socket-add-extra-source: write");
-				fflush(stderr);
-			} else {
-				printf("socket-add-extra-source: inserted additional message\n");
+				printf("socket-add-extra-source: inserting additional message\n");
 				fflush(stdout);
+
+				int len = strlen(message);
+				if (write(extra_fd, message, strlen(message)) != len) {
+					perror("socket-add-extra-source: write");
+					fflush(stderr);
+				} else {
+					printf("socket-add-extra-source: inserted additional message\n");
+					fflush(stdout);
+				}
+			} else {
+				fprintf(stderr, "socket-add-extra-source: no socket available\n");
+				fflush(stderr);
 			}
-		} else {
-			fprintf(stderr, "socket-add-extra-source: no socket available\n");
-			fflush(stderr);
 		}
+
+		active = false;
 	}
 
 	return (*next_execvp)(filename, argv);
